+++ pkgs/http/lib/src/mock_client.dart
@@ -7,6 +7,7 @@ import 'dart:convert';
 import 'base_client.dart';
 import 'base_request.dart';
 import 'byte_stream.dart';
+import 'progress.dart';
 import 'request.dart';
 import 'response.dart';
 import 'streamed_request.dart';
@@ -72,7 +73,10 @@ class MockClient extends BaseClient {
         });
 
   @override
-  Future<StreamedResponse> send(BaseRequest request) async {
+  Future<StreamedResponse> send(
+    BaseRequest request, {
+    Progress? onSendProgress,
+  }) async {
     var bodyStream = request.finalize();
     return await _handler(request, bodyStream);
   }